Add platform fru detection
Platform_scan on first run will instruct the fru_device to find
all available frus. It will then parse the entity json descriptions
and perform possible fru probes. When it finds a match, it will add
the entity into the gloabl fru structure. This then generates a system.json
file that can be used instead of having to parse again.

Change-Id: I3fab4cf1f0e7a7466bfc1be8d9f3fc2570da9ed0