Add a setting for AC Boot status
This adds a setting that defaults to "Unknown" and is updated to
"True" or "False" based on the boot status provided by u-boot on
the command line.

Tested:
Verified that the AC Boot setting is created with the default value
of "Unkown" on each boot.

Combined with the AC Boot check change, this makes the power
restore policy behave as expected.

Change-Id: I6358956464a71078fd8e3ff989a6fcf0b17b1028
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>