Add settings for Processor Error Config
This adds the settings needed for the Processor Error Config.

It stores the CATERR and ERR2 reset configuration and the current
count of CATERRs for each CPU and total crashdumps.

Tested:
Verified that the settings are available on DBus.

Combined with the IPMI change, verified that the Get and Set
Processor Error Config commands can get the data and update the
reset configuration both from IPMI directly and from BIOS setup.

Change-Id: I659ec31f59251c03cfd34a19fd0b1c5e20487d48
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>