Check for initial error state on startup
Since the error monitor is looking at GPIO edge events, an error
won't be detected if it is already asserted when the monintor
starts. This change detects if the host is already on when it
starts monitoring and will check if an error is already asserted.

Tested:
Injected an IERR while the error monitor was stopped. After
checking that the IERR pin was asserted, started the error monitor
and confirmed that the error was detected, logged, and handled
correctly.

Change-Id: Ie1e0238c914e6b2cda4b121f579776b813b191c1
Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>