Handle host errors on host power-on
Some errors such as the Boot IVR (Thermtrip) could be asserted
when the host powers-on.  This change initializes the error
status on host power-on to handle those types of errors.

Tested:
Moved the jumper to persistently assert Thermtrip and pressed the
power button.  After the system immediately shut down, confirmed
that the Thermtrip event was logged.

Change-Id: Id782167e234d906aaadb6190d9b92beafa6ab84a
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>