Allow raw PECI ping commands
The PECI ping command has a read length of zero, so when it's
called the response buffer is NULL since it has a size of zero.

This change allows the command to continue with a NULL response
buffer pointer as long as the read length is zero.

Tested:
Sent a raw ping command to a present CPU and got a successful
response.
Sent a raw ping command to a not-present CPU and got a failed
response.
Sent a raw read package config command and got the expected data.

Change-Id: I3e672b68e60716c8e239cab569d400f24760cb5b
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>