Separate stepping from model in peci_GetCPUID()
Including the stepping in the CPUID was too granular and was
causing a lot of unnecessary updates to code that uses it.

This change returns the CPU model as the main parameter for
identifying the processor and gives the stepping as a separate
parameter in case it is needed.

Tested:
Combined with the changes in the consumers of this library,
verified that the CPU model can correctly be retreived and
used to determine the correct PECI calls to use.

Change-Id: Ifef3ed4c952bf303bcb43aebb6a013e91719756e
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>