ref: Expose domain id as a string on dbus interface
So far domain ids where represented by number from predefined enum.
With latest NM change the domain ids will be exposed on dbus as strings
and numbers will be replaced with full domain names. 
So this change is an upgrade to these code changes.

Tested: Calling busctl tree phosphor-node-namager-proxy reveals that
the Domain is exposed with w full name.

Signed-off-by: Zbigniew Kurzynski <zbigniew.kurzynski@intel.com>
Change-Id: Ie6662cce36aabb99ef8cac5cd6eda29bbca98fed